configured environment to deploy map on apache server
