+++ map/FlaskApp.wsgi
@@ -1,0 +1,6 @@
+import sys
+import logging
+logging.basicConfig(stream=sys.stderr)
+sys.path.insert(0,"/var/www/diogenet_map/")
+
+from index import server as application
