--- map/app.py
@@ -1,7 +1,0 @@
-import dash
-import dash_bootstrap_components as dbc
-
-app = dash.Dash(__name__, external_stylesheets= [dbc.themes.BOOTSTRAP, dbc.icons.BOOTSTRAP])
-#app.config.suppress_callback_exceptions = True
-
-server = app.server
