--- map/data.csv
@@ -1,30 +1,0 @@
-lat,lon,text
-35.8888273,-106.3050223,Acid/Pueblo Canyon
-39.5034873,-84.7438586,Alba Craft Shop
-44.6208218,-123.1209165,"""Albany, Oregon, FUSRAP Site"""
-40.641371,-80.242936,Aliquippa Forge
-39.361063,-84.54075,Associated Aircraft Tool and Manufacturing Co.
-39.9573541,-83.0114552,B & T Metals
-41.672189,-83.568625,Baker Brothers
-40.746102,-74.0066415,Baker and Williams Warehouses
-35.8997249,-106.2901271,"""Bayo Canyon, New Mexico, FUSRAP Site"""
-42.841147,-78.83406,Bliss and Laughlin Steel
-41.8881016,-84.0164792,Bridgeport Brass Co.
-40.5376109,-79.7820937,C.H. Schnorr
-42.1542011,-72.4997776,Chapman Valve
-34.214928,-105.8793205,"""Chupadera Mesa, New Mexico, FUSRAP Site"""
-38.6870307,-90.1673616,"""Dow Chemical Co., Madison Site"""
-38.6926197,-90.16229,General Steel Industries
-37.872571,-122.256179,"""Gilman Hall, University of California at Berkeley"""
-39.3822485,-84.5471105,Hamilton OH Site
-40.7087363,-74.1012813,Jersey City FUSRAP Site
-40.577768,-74.4965644,Middlesex North Site
-41.7986843,-87.6069369,National Guard Armory
-40.4702797,-74.4762541,"""New Brunswick, NJ, FUSRAP Site"""
-43.2209056,-78.9737308,Niagara Falls Vicinity Properties
-36.0427296,-84.2066525,"""Oak Ridge, TN, Warehouse Site"""
-41.4007898,-73.072232,Seymour Specialty Wire
-42.9918367,-78.9201008,Tonawanda North FUSRAP Site
-41.7924797,-87.6010361,University of Chicago Metallurgical Laboratory
-42.5404703,-70.8888085,Ventron Corporation
-40.9696277,-74.2707793,Wayne FUSRAP Site
